// src/types/TariffBox.ts
export interface TariffBox {
	warehouseName: string
	geoName: string
	boxDeliveryBase: string
	boxDeliveryCoefExpr: string
	boxDeliveryLiter: string
	boxDeliveryMarketplaceBase: string
	boxDeliveryMarketplaceCoefExpr: string
	boxDeliveryMarketplaceLiter: string
	boxStorageBase: string
	boxStorageCoefExpr: string
	boxStorageLiter: string
}